requirements.txt
shiny>=0.6
anthropic
python-dotenv
